+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.1
 - Document `rdbg` compatibility.
+- Tweak the `vscode-rdbg` check so to enable `rdbg` support whenever that extension is not active.
 
 ## 0.1.0
 - Initial release.